Rename files for clarity, separate discrete randvar base from operations
