+++ tests/test_ops_discrete_randvars.py
